converted scripts needed to run examples in windows, added cifar100 dataset builder and a loss/accuracy plotter
This is the updated version of my previous branch which tried to convert
existing examples to windows, added a new loss/acurracy plotter script .

in  this update, the linux based scripts are included and some scripts
are fixed as well.
I also added a 3dparty folder (gnu tools) which consists of apps that
mimic linux apps and commands to reduce the amount of changes to the
script.